{{cookiecutter.project_slug}}/backend/app/pyproject.toml
[tool.poetry]
name = "nexly"
version = "0.0.1"
description = ""
authors = ["Liakis Stefanos <stliakis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.1"
fastapi = "^0.68.1"
python-multipart = "^0.0.5"
requests = "^2.23.0"
celery = "^4.4.2"
passlib = { extras = ["bcrypt"], version = "^1.7.2" }
tenacity = "^6.2.0"
pydantic = "^1.4"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^2.11.2"
psycopg2-binary = "^2.8.5"
alembic = "^1.4.2"
sqlalchemy = "^1.3.16"
pytest = "^5.4.1"
elasticsearch = "^7.13.4"
deepmerge = "^0.3.0"
python-jose = { extras = ["cryptography"], version = "^3.1.0" }
redis = "^3.5.3"
pymemcache = "^3.5.0"
fastapi-utils = "^0.2.1"
httpcore = "^0.15"
pymongo = "^4.4.1"
jurigged = { extras = ["develoop"], version = "^0.4.1" }
ipdb = "^0.13.9"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
sqlalchemy-stubs = "^0.3"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

